post refactoring tweaks; changes to catching ImportError
