+++ nltk/twitter/__init__.py
@@ -13,14 +13,15 @@ This package contains classes for retrieving Tweet documents using the
 Twitter API.
 
 """
-try:
-    from twython import Twython, TwythonStreamer
-except ImportError:
-    raise ImportError("""The twitterclient module requires the Twython
-    package. See https://twython.readthedocs.org/ for installation
-    instructions.""")
+#try:
+    #from twython import Twython, TwythonStreamer
+#except ImportError as e:
+    #e.msg = """The twitterclient module requires the Twython
+    #package. See https://twython.readthedocs.org/ for installation
+    #instructions."""
+    #raise
 
-from nltk.util import credsfromfile
+from nltk.twitter.util import credsfromfile
 from nltk.twitter.twitterclient import Streamer, Query, Twitter, TweetViewer,\
      TweetWriter
 
@@ -29,3 +30,7 @@ def setup_module(module):
     from nose import SkipTest
     raise SkipTest("nltk.twitter examples are not doctests")
 
+
+if __name__ == "__main__":
+    from twitterclient import Twitter
+    tw = Twitter()
